[storybook] Move <DebugBox> & <FlexRow> to utils/
Lets webpack can resolve the modules in storybook/ package.
