+++ packages/storybook/examples/TextLabel/BasicTextLabel.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import TextLabel from '@ichef/gypcrete/src/TextLabel';
-import DebugBox from '../DebugBox';
+import DebugBox from 'utils/DebugBox';
 
 function BasicTextLabelExample() {
     return (
